Pulse_Backend/backend/requirements.txt
cycler==0.10.0
Django==2.2.7
django-cors-headers==3.2.0
django-rest-framework==0.1.0
djangorestframework==3.10.3
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.4
Pillow==6.2.1
psycopg2==2.8.4
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
wordcloud==1.5.0
